refactor: isolate `EthrStatusRegistry` into a file
