+++ src/devTypes.d.ts
@@ -1,0 +1,3 @@
+declare module 'ethjs-signer'
+declare module 'ganache-cli'
+declare module '@truffle/contract'
