+++ src/types.d.ts
@@ -1,3 +1,5 @@
 declare module 'ethjs-provider-http'
 declare module 'ethjs-query'
 declare module 'ethjs-contract'
+
+declare module 'ethjs-provider-signer'
