+++ tsconfig.json
@@ -7,7 +7,8 @@
     "sourceMap": true,
     "outDir": "lib",
     "strict": true,
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "esModuleInterop": true
   },
   "include": [
     "src"
